refactor graphein.protein.utils.is_tool
